Split codegen file.
