+++ src/tiro/codegen/CMakeLists.txt
@@ -4,10 +4,12 @@ target_sources(tiro_objects
         basic_block.hpp
         code_builder.cpp
         code_builder.hpp
-        codegen.cpp
-        codegen.hpp
         expr_codegen.cpp
         expr_codegen.hpp
+        func_codegen.cpp
+        func_codegen.hpp
+        module_codegen.cpp
+        module_codegen.hpp
         stmt_codegen.cpp
         stmt_codegen.hpp
         variable_locations.cpp
