+++ src/tiro/codegen/expr_codegen.hpp
@@ -1,20 +1,25 @@
 #ifndef TIRO_CODEGEN_EXPR_CODEGEN_HPP
 #define TIRO_CODEGEN_EXPR_CODEGEN_HPP
 
-#include "tiro/codegen/codegen.hpp"
+#include "tiro/codegen/code_builder.hpp"
 #include "tiro/core/function_ref.hpp"
+#include "tiro/core/not_null.hpp"
+#include "tiro/syntax/ast.hpp"
 
 namespace tiro::compiler {
+
+class FunctionCodegen;
+class ModuleCodegen;
 
 /// This class is responsible for compiling expressions to bytecode.
 class ExprCodegen final {
 public:
-    ExprCodegen(NotNull<Expr*> expr, FunctionCodegen& func);
+    explicit ExprCodegen(NotNull<Expr*> expr, FunctionCodegen& func);
 
     /// Returns false if value generation was omitted as an optimization.
     bool generate();
 
-    ModuleCodegen& module() { return func_.module(); }
+    ModuleCodegen& module();
 
 public:
     bool visit_unary_expr(UnaryExpr* e);
